Remove total raw walltime (#120)
The total raw walltime is (was!) the total time taken by a region with
the overhead time from calling direct children removed. And, crucially,
*only* direct children. It's tempting to think that it might be the time
with overheads from the downward call tree removed, but it isn't.

My thinking when I originally put it in was around getting a handle on
where in the code the overhead was coming from. At the time, it seemed
sensible to create this thing called raw walltime and output that. There
are two problems with that:

* In practical terms, so I discovered, it's easier just to have the
overhead time
* It leads to confusion about what the raw total time represents. I
certainly got confused.

There are also complications around the raw time with vs. without
recursions.

I wondered about putting in a mechanism to compute true raw total
walltimes, being an estimate of the total time we would have got if
there was no profiling overhead. That requires calculating the overhead
accrued in whole call trees. I attempted to code this, and it got very
messy. So, best to stick with total times including the overhead time, I
think.

This PR reverts PR #107. Sorry - I hadn't thought things through enough
when I opened #104.