Add support for single file output mode
This adds support for a single global output file using MPI IO to
create a file view for each task and a collective write to populate
the file.

There are more than few rough edges - including a possible memory leak
in the output code.

The Writer class also feels like it needs a refactor because it is not
clear that the output stream needs to be exposed to the outside world.
Dealing with this would also remove a warning about an unused argument
in the SingleFile.write member function.